requirements.txt
spotipy
python-dotenv
tqdm